Algo: Replace Space
Write a method to replace all spaces in a string with '%20'. You may
assume that the string has sufficient space at the end to hold
additional characters, and that you are given the "true" length of the
string. (Note: if implementing in Java, please use a character array so
that you can perform this operation in place.)
Example
Input: "Mr John Smith    ", 13
Output: "Mr%20John%20Smith"